eccr-cms/requirements.txt
boto3~=1.40.67

coverage~=7.12.0

ddt>=1.4.2, <1.5.0

Django>=4.2,<5.0

django-admin-interface>=0.28.3

django-colorfield>=0.11.0

django-cors-headers>=3.7.0, <3.8.0

django-csp>=3.7,<3.8

django-model-utils>=4.3.1,<5.0.0

djangorestframework>=3.15.2,<3.16.0

flake8>=3.9.0,<4.0

gunicorn>=23.0.0,<23.1.0

mysql-connector-python >= 9.5.0, <9.6.0

p1-auth>=0.1.6,<1.0.0

# Pillow>=10.3.0, <10.4.0

pipdeptree==2.30.0

psycopg2-binary >= 2.9, < 3.0

pylint-django == 2.5.3

requests~=2.32

setuptools==80.9.0

tzdata==2025.3